--- spark/dl/src/main/scala/com/intel/analytics/bigdl/example/int8/README.md
@@ -1,1 +1,0 @@
-#
